requirements.txt
fastapi
uvicorn
python-dotenv
langchain
langchain-openai
langchain-community
langchain-core
langchain-text-splitters
langchain-chroma
chromadb
